[Feat.] manual checks update
