+++ src/components/Table.module.css
@@ -1,0 +1,4 @@
+.wordBreak {
+    word-break: normal;
+    overflow-wrap: normal;
+}
